move ComponentShowcaseDialog button to ComponentShowcaseTab
